aggregate all log messages in one file & fix imports (#162)
* aggregate all log messages in one file & fix imports

* fix tests
