--- src/errors/generalError.ts
@@ -1,3 +1,0 @@
-import createError from 'fastify-error'
-
-export const GeneralError = createError('ERR_GENERAL_ERROR', 'Error', 500)
