--- src/errors/invalidConfig.ts
@@ -1,3 +1,0 @@
-import createError from 'fastify-error'
-
-export const InvalidConfigError = createError('ERR_INVALID_CONFIG', `[FATAL] Invalid config.`, 500)
