--- src/errors/invalidUrl.ts
@@ -1,3 +1,0 @@
-import createError from 'fastify-error'
-
-export const InvalidUrl = createError('ERR_INVALID_URL', `Invalid Url.`, 400)
