--- src/errors/unauthorized.ts
@@ -1,3 +1,0 @@
-import createError from 'fastify-error'
-
-export const UnauthorizedError = createError('ERR_UNAUTHORIZED', 'Unauthorized', 401)
