+++ src/services/auth/drivers/bearerToken/index.ts
@@ -1,8 +1,7 @@
 import { AuthDriver } from '../../types'
 import { BearerTokenDriverConfig } from './types'
 import { FastifyRequest } from 'fastify'
-import { UnauthorizedError } from '../../../../errors/unauthorized.js'
-import { logger } from '../../../logger/logger.js'
+import { UnauthorizedError } from '../../../../errors/errors.js'
 
 export class BearerTokenAuth implements AuthDriver {
 	private token: string
