+++ src/services/auth/index.ts
@@ -1,5 +1,5 @@
 import { FastifyRequest } from 'fastify'
-import { InvalidConfigError } from '../../errors/invalidConfig.js'
+import { InvalidConfigError } from '../../errors/errors.js'
 import { BearerTokenAuth } from './drivers/bearerToken/index.js'
 import type { AuthDriver } from './types'
 import { AuthConfig, AuthDriverName } from './types/config.js'
