+++ src/services/storage/drivers/inMemory/index.ts
@@ -1,6 +1,5 @@
 import cryptoRandomString from 'crypto-random-string'
-import { GeneralError } from '../../../../errors/generalError.js'
-import { NotFoundError } from '../../../../errors/notFound.js'
+import { NotFoundError , GeneralError } from '../../../../errors/errors.js'
 import { InMemoryStorageConfig } from '../../types/config.js'
 import type { StorageDriver } from '../../types/index.js'
 import type { StoredUrl, UrlWithInformation, UrlRequestData, UrlInformation } from '../../types/url.js'
