+++ src/services/urlValidator.ts
@@ -1,6 +1,6 @@
 import { config } from '../config/index.js'
 import micromatch from 'micromatch'
-import { InvalidUrl } from '../errors/invalidUrl.js'
+import { InvalidUrl } from '../errors/errors.js'
 
 const matcher = micromatch.matcher(config.url.matchPattern)
 export async function validateUrl(url: string): Promise<void> {
